fix -file path

+++ src/apps/shared/models/scraper_data.py
@@ -8,6 +8,11 @@ from bs4 import BeautifulSoup
 from pymongo import MongoClient
 from datetime import datetime
 import gridfs
+import os
+
+output_dir = r"C:\web_scraping_files"
+if not os.path.exists(output_dir):
+    os.makedirs(output_dir)
 
 driver = webdriver.Chrome(service=Service(ChromeDriverManager().install()))
 
@@ -30,31 +35,36 @@ try:
     )
 
     soup = BeautifulSoup(driver.page_source, 'html.parser')
-
     text_content = soup.get_text(separator='\n', strip=True)
 
-    file_name = 'iucngisd.txt'
-    
-    with open(file_name, 'w', encoding='utf-8') as file:
-        file.write(text_content)  
-    
-    with open(file_name, 'rb') as file_data:
-        object_id = fs.put(file_data, filename=file_name)  
+    existing_doc = collection.find_one({"Url": 'http://www.iucngisd.org/gisd/'})
 
-    data = {
-        'Objeto': object_id,  
-        'Tipo': 'Web', 
-        'Url': 'http://www.iucngisd.org/gisd/', 
-        'Fecha': datetime.now().strftime('%Y-%m-%d %H:%M:%S'),  
-        'Etiquetas': ['planta', 'plaga'], 
-    }
+    if existing_doc:
+        print(f"Ya existe un documento para esta URL con ObjectId: {existing_doc['Objeto']}")
+    else:
+        file_name = os.path.join(output_dir, 'iucngisd.txt')
+        
+        with open(file_name, 'w', encoding='utf-8') as file:
+            file.write(text_content)
+        
+        with open(file_name, 'rb') as file_data:
+            object_id = fs.put(file_data, filename=file_name)  
 
-    collection.insert_one(data)
+        data = {
+            'Objeto': object_id,  
+            'Tipo': 'Web', 
+            'Url': 'http://www.iucngisd.org/gisd/', 
+            'Fecha_scrapper': datetime.now().strftime('%Y-%m-%d %H:%M:%S'),  
+            'Etiquetas': ['planta', 'plaga'], 
+        }
 
-    print('Los datos se han guardado en MongoDB y el contenido se ha escrito en el archivo.')
+        collection.insert_one(data)
+
+        print(f"Los datos se han guardado en MongoDB y el contenido se ha escrito en el archivo. ObjectId: {object_id}")
 
 except Exception as e:
     print(f'Ocurrió un error: {e}')
 
 finally:
+    # Cerrar el navegador
     driver.quit()
